+++ testbench_ai_service/main.py
@@ -12,7 +12,10 @@ from testbench_ai_service.agents.routes import get_agent_routers
 from testbench_ai_service.config import AppConfig
 from testbench_ai_service.exceptions import http_exception_handler
 from testbench_ai_service.llm.factory import LLMFactory
-from testbench_ai_service.middlewares import LoggingMiddleware, OutboundRequestLoggingMiddleware
+from testbench_ai_service.middlewares import (
+    LoggingMiddleware,
+    OutboundRequestLoggingMiddleware,
+)
 from testbench_ai_service.routes import router
 from testbench_ai_service.utils.config import load_config_from_file
 from testbench_ai_service.utils.i18n import load_translations
@@ -54,7 +57,10 @@ def init_exception_handlers(app: FastAPI):
 def init_middlewares(app: FastAPI):
     """Initialization of app middlewares"""
 
-    app.add_middleware(LoggingMiddleware)
+    app.add_middleware(
+        LoggingMiddleware,
+        max_payload_length=app.state.config.logging.max_payload_length,
+    )
     OutboundRequestLoggingMiddleware.install(app.state.config.tb_server_url)